Change _ConvertLineToUserObject to use stoi for integers
Updated the `_ConvertLineToUserObject` method in the `clsUser`
class to use `stoi` instead of `stof` for the sixth parameter
conversion. This reflects a change in the expected data type
from float to integer for the sixth field in the `vUserData`
vector, aligning with the updated `clsUser` constructor
requirements.

+++ clsUser.h
@@ -46,7 +46,7 @@ private:
 	{
 		vector<string> vUserData = clsString::Split(Line, Separator);
 
-		return clsUser(enMode::UpdateMode, vUserData[0], vUserData[1], vUserData[2], vUserData[3], vUserData[4], vUserData[5], stof(vUserData[6]));
+		return clsUser(enMode::UpdateMode, vUserData[0], vUserData[1], vUserData[2], vUserData[3], vUserData[4], vUserData[5], stoi(vUserData[6]));
 	}
 
 	static vector<clsUser> _LoadUsersDataFromFile(string FileName = "Users.txt")
